requirements.txt
telethon==1.29.1
google-api-python-client==2.121.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
